Replace long by int to have compatibility with external libraries

+++ Html5/Gamepad.cs
@@ -12,7 +12,7 @@
         /// <summary>
         /// The index of the gamepad in the Navigator. When multiple gamepads are connected to a user agent, indices must be assigned on a first-come, first-serve basis, starting at zero. If a gamepad is disconnected, previously assigned indices must not be reassigned to gamepads that continue to be connected. However, if a gamepad is disconnected, and subsequently the same or a different gamepad is then connected, index entries must be reused.
         /// </summary>
-        public readonly long Index;
+        public readonly int Index;
 
         /// <summary>
         /// Indicates whether the physical device represented by this object is still connected to the system. When a gamepad becomes unavailable, whether by being physically disconnected, powered off or otherwise unusable, the connected attribute must be set to false.
